Added cross-compilem toolchain selection support. common built with gcc 4.4.1 (arm-2009q3 toolchain)
<==== Mercurial data ====>
Branch: qt5port

+++ common/src/plugins/resources/axis/axis_ptz_controller.cpp
@@ -115,13 +115,13 @@ void QnAxisPtzController::init(const QnAxisParameterMap &params) {
 
         /* Axis takes care of image rotation automagically, but we still need to adjust tilt limits. */
         qreal rotation = params.value("root.Image.I0.Appearance.Rotation", 0.0);
-        if(qFuzzyCompare(rotation, 180.0)) {
+        if(qFuzzyCompare(rotation, (qreal)180.0)) {
             minTilt = -minTilt;
             maxTilt = -maxTilt;
             qSwap(minTilt, maxTilt);
         }
 
-        QnScalarSpaceMapper xMapper(minPan, maxPan, minPan, maxPan, qFuzzyCompare(maxPan - minPan, 360.0) ? Qn::PeriodicExtrapolation : Qn::ConstantExtrapolation);
+        QnScalarSpaceMapper xMapper(minPan, maxPan, minPan, maxPan, qFuzzyCompare(maxPan - minPan, (qreal)360.0) ? Qn::PeriodicExtrapolation : Qn::ConstantExtrapolation);
         QnScalarSpaceMapper yMapper(minTilt, maxTilt, minTilt, maxTilt, Qn::ConstantExtrapolation);
         QnScalarSpaceMapper zMapper(1, 9999, fovTo35mmEquiv(gradToRad(maxAngle)), fovTo35mmEquiv(gradToRad(minAngle)), Qn::ConstantExtrapolation);
         /* Note that we do not care about actual zoom limits on the camera.
